Bluetooth: Remove notification once device is bonded
This change removes the pairing notification once the
bonding process is successfully completed. This scenario
is specific to the cases where external apps set the Pin
Key while bonding and the pairing popups or the notifications
that are generated by the Settings apk need to be removed
at the end of successful pairing or a failed bond.

Change-Id: Iee62ca36ae4cc61dce70e2fd9c85959a68d7664a